feat: make use of Jekyll for a cleaner site
I did this for a number of reasons that I won't go into, but the bigger
concern I have about this change is whether or not it will work properly
with GitHub Pages.

I know GitHub Pages has native support for Jekyll, but I have no idea if
pushing this repo in as is will actually work properly for the site.

Only one way to find out.