refactor: stop tracking _site/
Apparently GitHub Pages will ignore it in the repository and just build
itself, so I might as well not push it to the repo.

There are also some miscellaneous changes here to the footer and the
contact pages; I know it's not best practice to include these changes in
the same commit, but idc since it's my small little project anyway.